move dictionary to separate folder
